+++ src/routes/docs/auth.md
