+++ src/routes/docs/auth.replace.md
