+++ src/routes/docs/rooms.md
