+++ src/routes/rooms.js
